gateway/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
aiokafka==0.10.0
redis
kafka-python